Move Device model to root and re-arrange Product
