+++ tests/test_helpers.py
@@ -6,7 +6,7 @@ from unittest.mock import AsyncMock
 import pytest
 from aiohttp import ClientSession
 
-from homewizard_energy import get_verification_hostname, has_v2_api
+from homewizard_energy import has_v2_api
 
 pytestmark = [pytest.mark.asyncio]
 
@@ -66,31 +66,3 @@ async def test_has_v2_api_own_session(aresponses):
 
     result = await has_v2_api("example.com")
     assert result is True
-
-
-@pytest.mark.parametrize(
-    ("model", "expected"),
-    [
-        ("HWE-P1", "appliance/p1dongle/1234567890"),
-        ("HWE-SKT", "appliance/energysocket/1234567890"),
-        ("HWE-WTR", "appliance/watermeter/1234567890"),
-        ("HWE-DSP", "appliance/display/1234567890"),
-        ("HWE-KWH1", "appliance/energymeter/1234567890"),
-        ("SDM230-wifi", "appliance/energymeter/1234567890"),
-        ("HWE-KWH3", "appliance/energymeter/1234567890"),
-        ("SDM630-wifi", "appliance/energymeter/1234567890"),
-        ("HWE-BAT", "appliance/battery/1234567890"),
-    ],
-)
-async def test_get_verification_hostname_returns_expected_identifier(
-    model: str, expected: str
-):
-    """Test if get_verification_hostname returns the expected identifier."""
-    result = get_verification_hostname(model, "1234567890")
-    assert result == expected
-
-
-def test_get_verification_hostname_raises_value_error():
-    """Test if get_verification_hostname raises a ValueError for an unsupported model."""
-    with pytest.raises(ValueError):
-        get_verification_hostname("unsupported", "1234567890")
