+++ tests/v2/test_v2_models.py
@@ -5,7 +5,8 @@ import json
 import pytest
 from syrupy.assertion import SnapshotAssertion
 
-from homewizard_energy.v2 import Device, Measurement, System, SystemUpdate
+from homewizard_energy.models import Device
+from homewizard_energy.v2.models import Measurement, System, SystemUpdate
 
 from . import load_fixtures
